backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==43.0.3
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
httpx==0.28.1
fastembed==0.6.0
onnxruntime==1.20.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.1
pytest==8.3.4
python-dotenv==1.2.1
python-multipart==0.0.20
python-jose==3.5.0
pypdf==6.0.0
python-docx==1.2.0
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.1.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.36
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
